+++ internal/service/internal/handler/config_connects_test.go
@@ -2,15 +2,169 @@ package handler
 
 import (
 	"context"
+	"errors"
 	"fmt"
+	"io"
 	"net/http"
+	"net/http/httptest"
+	"strings"
 	"testing"
 	"time"
 
+	"github.com/number571/go-peer/pkg/logger"
 	"github.com/number571/hidden-lake/internal/service/pkg/app/config"
 	hls_client "github.com/number571/hidden-lake/internal/service/pkg/client"
+	std_logger "github.com/number571/hidden-lake/internal/utils/logger/std"
 	testutils "github.com/number571/hidden-lake/test/utils"
 )
+
+func TestHandleConnectsAPI2(t *testing.T) {
+	t.Parallel()
+
+	ctx := context.Background()
+
+	httpLogger := std_logger.NewStdLogger(
+		func() std_logger.ILogging {
+			logging, err := std_logger.LoadLogging([]string{})
+			if err != nil {
+				panic(err)
+			}
+			return logging
+		}(),
+		func(_ logger.ILogArg) string {
+			return ""
+		},
+	)
+
+	node := newTsNode()
+	handler := HandleConfigConnectsAPI(ctx, newTsWrapper(true), httpLogger, node)
+	if err := connectsAPIRequestOK(handler); err != nil {
+		t.Error(err)
+		return
+	}
+	if err := connectsAPIRequestPostOK(handler); err != nil {
+		t.Error(err)
+		return
+	}
+	if err := connectsAPIRequestDeleteOK(handler); err != nil {
+		t.Error(err)
+		return
+	}
+
+	if err := connectsAPIRequestMethod(handler); err == nil {
+		t.Error("request success with invalid method")
+		return
+	}
+	if err := connectsAPIRequestPostConnect(handler); err == nil {
+		t.Error("request success with invalid connect")
+		return
+	}
+
+	handlerx := HandleConfigConnectsAPI(ctx, newTsWrapper(false), httpLogger, node)
+	if err := connectsAPIRequestPostOK(handlerx); err == nil {
+		t.Error("request success with invalid update editor (post)")
+		return
+	}
+	if err := connectsAPIRequestDeleteOK(handlerx); err == nil {
+		t.Error("request success with invalid update editor (delete)")
+		return
+	}
+}
+
+func connectsAPIRequestDeleteOK(handler http.HandlerFunc) error {
+	w := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodDelete, "/", strings.NewReader("127.0.0.1:9999"))
+
+	handler(w, req)
+	res := w.Result()
+	defer res.Body.Close()
+
+	if res.StatusCode != http.StatusOK {
+		return errors.New("bad status code") // nolint: err113
+	}
+
+	if _, err := io.ReadAll(res.Body); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func connectsAPIRequestPostOK(handler http.HandlerFunc) error {
+	w := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodPost, "/", strings.NewReader("127.0.0.1:9999"))
+
+	handler(w, req)
+	res := w.Result()
+	defer res.Body.Close()
+
+	if res.StatusCode != http.StatusOK {
+		return errors.New("bad status code") // nolint: err113
+	}
+
+	if _, err := io.ReadAll(res.Body); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func connectsAPIRequestOK(handler http.HandlerFunc) error {
+	w := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodGet, "/", nil)
+
+	handler(w, req)
+	res := w.Result()
+	defer res.Body.Close()
+
+	if res.StatusCode != http.StatusOK {
+		return errors.New("bad status code") // nolint: err113
+	}
+
+	if _, err := io.ReadAll(res.Body); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func connectsAPIRequestPostConnect(handler http.HandlerFunc) error {
+	w := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodPost, "/", strings.NewReader(""))
+
+	handler(w, req)
+	res := w.Result()
+	defer res.Body.Close()
+
+	if res.StatusCode != http.StatusOK {
+		return errors.New("bad status code") // nolint: err113
+	}
+
+	if _, err := io.ReadAll(res.Body); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func connectsAPIRequestMethod(handler http.HandlerFunc) error {
+	w := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodPut, "/", nil)
+
+	handler(w, req)
+	res := w.Result()
+	defer res.Body.Close()
+
+	if res.StatusCode != http.StatusOK {
+		return errors.New("bad status code") // nolint: err113
+	}
+
+	if _, err := io.ReadAll(res.Body); err != nil {
+		return err
+	}
+
+	return nil
+}
 
 func TestHandleConnectsAPI(t *testing.T) {
 	t.Parallel()